image-data/requirements.txt
transformers
torch
packaging
wheel
flash_attn
pillow
timm
accelerate